app/requirements.txt
pydub==0.24.1
PyAudio==0.2.11
git+https://github.com/acrcloud/acrcloud_sdk_python
gTTS==2.1.1
smbus2==0.2.1
requests==2.24.0
ipython
RPi.GPIO
